# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  http.NodeResponse:
    properties:
      besuNode:
        $ref: '#/definitions/service.BesuNodeProperties'
      createdAt:
        type: string
      endpoint:
        type: string
      fabricOrderer:
        $ref: '#/definitions/service.FabricOrdererProperties'
      fabricPeer:
        $ref: '#/definitions/service.FabricPeerProperties'
      id:
        type: integer
      name:
        type: string
      nodeType:
        type: string
      platform:
        type: string
      status:
        type: string
      updatedAt:
        type: string
    type: object
...
  http.PaginatedNodesResponse:
    properties:
      hasNextPage:
        type: boolean
      items:
        items:
          $ref: '#/definitions/http.NodeResponse'
        type: array
      page:
        type: integer
      pageCount:
        type: integer
      total:
        type: integer
    type: object
...
  response.ErrorResponse:
    properties:
      details:
        additionalProperties: true
        type: object
      message:
        type: string
      type:
        type: string
    type: object
...
host: localhost:8100
info:
  contact:
    email: support@chainlaunch.com
    name: API Support
    url: http://www.chainlaunch.com/support
  description: ChainLaunch API provides services for managing blockchain networks
    and cryptographic keys
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: ChainLaunch API
  version: "1.0"
paths:
...
  /nodes:
    get:
      consumes:
      - application/json
      description: Get a paginated list of nodes with optional platform filter
      parameters:
      - description: Filter by blockchain platform
        in: query
        name: platform
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.PaginatedNodesResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: List all nodes
      tags:
      - nodes
    post:
      consumes:
      - application/json
      description: Create a new node with the specified configuration
      parameters:
      - description: Node creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.CreateNodeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http.NodeResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Create a new node
      tags:
      - nodes
  /nodes/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a node by ID
      parameters:
      - description: Node ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Node not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Delete a node
      tags:
      - nodes
    get:
      consumes:
      - application/json
      description: Get a node by ID
      parameters:
      - description: Node ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.NodeResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Node not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get a node
      tags:
      - nodes
  /nodes/{id}/events:
    get:
      consumes:
      - application/json
      description: Get a paginated list of events for a specific node
      parameters:
      - description: Node ID
        in: path
        name: id
        required: true
        type: integer
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.PaginatedNodeEventsResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Node not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get node events
      tags:
      - nodes
  /nodes/{id}/logs:
    get:
      consumes:
      - application/json
      description: Stream logs from a specific node
      parameters:
      - description: Node ID
        in: path
        name: id
        required: true
        type: integer
      - default: false
        description: Follow logs
        in: query
        name: follow
        type: boolean
      - default: 100
        description: Number of lines to show from the end
        in: query
        name: tail
        type: integer
      produces:
      - text/event-stream
      responses:
        "200":
          description: Log stream
          schema:
            type: string
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Node not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Tail node logs
      tags:
      - nodes
  /nodes/{id}/restart:
    post:
      consumes:
      - application/json
      description: Restart a node by ID (stops and starts the node)
      parameters:
      - description: Node ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.NodeResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Node not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Restart a node
      tags:
      - nodes
  /nodes/{id}/start:
    post:
      consumes:
      - application/json
      description: Start a node by ID
      parameters:
      - description: Node ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.NodeResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Node not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Start a node
      tags:
      - nodes
  /nodes/{id}/stop:
    post:
      consumes:
      - application/json
      description: Stop a node by ID
      parameters:
      - description: Node ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.NodeResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Node not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Stop a node
      tags:
      - nodes
  /nodes/defaults/besu-node:
    get:
      description: Get default configuration values for a Besu node
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.BesuNodeDefaults'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get default values for Besu node
      tags:
      - nodes
  /nodes/defaults/fabric:
    get:
      description: Get default configuration values for multiple Fabric nodes
      parameters:
      - default: 1
        description: Number of peer nodes
        in: query
        minimum: 0
        name: peerCount
        type: integer
      - default: 1
        description: Number of orderer nodes
        in: query
        minimum: 0
        name: ordererCount
        type: integer
      - default: service
        description: Deployment mode
        enum:
        - service
        - docker
        in: query
        name: mode
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.NodesDefaultsResult'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get default values for multiple Fabric nodes
      tags:
      - nodes
  /nodes/defaults/fabric-orderer:
    get:
      description: Get default configuration values for a Fabric orderer node
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.NodeDefaults'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get default values for Fabric orderer node
      tags:
      - nodes
...
schemes:
- http
- https
securityDefinitions:
  BasicAuth:
    type: basic
  CookieAuth:
    in: cookie
    name: session_id
    type: apiKey
swagger: "2.0"
...